Import the column the writer writes, and stop copying volumes to read them
Three things the audit turned up, none of them dead code - there is none: no
function or constant in the page or in src/ is unreferenced.

geometry_to_blender.py has been broken since the surface kinds landed on Sep 4.
It asks parquet for a `mesh` column that was renamed to `surface`, then filters
on `df["surface"]`, which it never requested - so it fails on every geometry a
current run writes. Past that it would have crashed anyway: decode_mesh reads the
first eight bytes of whatever it is handed as nV and nF, and an ellipsoid is 60
bytes that read as nV = 1.09 billion, which numpy answers with "buffer is smaller
than requested size" from inside the merge. The tests never saw either: they hand
decode_mesh a payload directly, and the read is in main(), which needs Blender.

So: the right columns, a length check derived from the header (a payload of
another kind is not a short mesh, it is not a mesh), and the parametric kinds
counted and left out with the flag that would have stored them as meshes instead.
The column names are now checked against a geometry file a real run writes, which
is the test that was missing.

Two volumes were being copied to be read:

  find_objects(labels.astype(np.int32))     0.60 s, peak +562 MB
  find_objects(labels)                      0.42 s, peak +  1 MB

find_objects takes any integer type, and entities are stacked in the narrowest
one their ids need. Contacts is where a run peaks and that peak is what sizes the
object pool, so this was the worst place to hold a second copy of the largest
array in it.

  astype + two `> 0` passes                 0.44 s, peak +702 MB
  one `> 0`, reused                         0.17 s, peak +141 MB

Same on the entity row: an int32 copy of the volume, then `> 0` twice - once to
select the ids and once to count them. Same answers from one pass.

Measured on a 300x700x700 uint16 entity, which is one alpha cell's granules.

+++ src/label_anatomy/analysis/gaps.py
@@ -69,7 +69,11 @@ def _gaps_within(
     # Bounding-box padding per axis, so everything within max_gap is inside the window. One
     # entry per spatial axis, which is the only difference between 2D and 3D here.
     reach = [int(math.ceil(max_gap_um / v)) for v in anisotropy]
-    boxes = find_objects(labels.astype(np.int32, copy=False))
+    # Not cast to int32 first: find_objects takes any integer type, and entities are
+    # stacked in the narrowest one their ids need (usually uint16), so the cast was a
+    # full second copy of the largest array this measurer holds - and contacts are where
+    # a run peaks, which is what sizes the object pool.
+    boxes = find_objects(labels)
 
     closest: Dict[Tuple[int, int], float] = {}
     for label_id, box in enumerate(boxes, start=1):
